# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /journals:
    get:
      operationId: listJournals
      tags: [journals]
      summary: List journal entries
      description: Returns the authenticated user's journal entries, optionally scoped to a month, for the calendar/history view.
      parameters:
        - name: month
          in: query
          schema: { type: string, pattern: "^\\d{4}-\\d{2}$" }
          description: "Restrict results to this calendar month, format YYYY-MM."
      responses:
        "200":
          description: Journal entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JournalEntry"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /journals/{date}:
    get:
      operationId: getJournal
      tags: [journals]
      summary: Get the journal entry for a date
      parameters:
        - name: date
          in: path
          required: true
          schema: { type: string, pattern: "^\\d{4}-\\d{2}-\\d{2}$" }
      responses:
        "200":
          description: Journal entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
        "404":
          description: No entry for this date
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: upsertJournal
      tags: [journals]
      summary: Create, update, or autosave the journal entry for a date
      parameters:
        - name: date
          in: path
          required: true
          schema: { type: string, pattern: "^\\d{4}-\\d{2}-\\d{2}$" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JournalInput"
      responses:
        "200":
          description: Journal entry saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteJournal
      tags: [journals]
      summary: Delete the journal entry for a date
      parameters:
        - name: date
          in: path
          required: true
          schema: { type: string, pattern: "^\\d{4}-\\d{2}-\\d{2}$" }
      responses:
        "204":
          description: Deleted
        "404":
          description: No entry for this date
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /rules:
    get:
      operationId: listRules
      tags: [rules]
      summary: List trading rules
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rule"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createRule
      tags: [rules]
      summary: Create a trading rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RuleInput"
      responses:
        "201":
          description: Rule created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /rules/{id}:
    patch:
      operationId: updateRule
      tags: [rules]
      summary: Update or toggle a trading rule
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RuleUpdate"
      responses:
        "200":
          description: Rule updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteRule
      tags: [rules]
      summary: Delete a trading rule
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      description: |
        Returns a presigned GCS URL for direct upload. The client sends JSON
        metadata here, then uploads the file directly to the returned URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity from PRIVATE_OBJECT_DIR
      parameters:
        - name: objectPath
          in: path
          required: true
          schema: { type: string }
          description: "Object path within the private object dir (e.g. `uploads/some-uuid`)."
      responses:
        "200":
          description: Object content streamed with correct Content-Type.
          content:
            "*/*":
              schema: { type: string, format: binary }
        "404":
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Failed to serve object.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    JournalEntry:
      type: object
      required:
        - id
        - date
        - mood
        - confidence
        - discipline
        - fear
        - greed
        - focus
        - sleep
        - tradingPlan
        - notes
        - mistakes
        - lessons
        - tomorrowGoal
        - isDraft
        - createdAt
        - updatedAt
      properties:
        id: { type: integer }
        date: { type: string }
        mood: { type: ["integer", "null"] }
        confidence: { type: ["integer", "null"] }
        discipline: { type: ["integer", "null"] }
        fear: { type: ["integer", "null"] }
        greed: { type: ["integer", "null"] }
        focus: { type: ["integer", "null"] }
        sleep: { type: ["integer", "null"] }
        tradingPlan: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        mistakes: { type: ["string", "null"] }
        lessons: { type: ["string", "null"] }
        tomorrowGoal: { type: ["string", "null"] }
        isDraft: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    JournalInput:
      type: object
      properties:
        mood: { type: ["integer", "null"] }
        confidence: { type: ["integer", "null"] }
        discipline: { type: ["integer", "null"] }
        fear: { type: ["integer", "null"] }
        greed: { type: ["integer", "null"] }
        focus: { type: ["integer", "null"] }
        sleep: { type: ["integer", "null"] }
        tradingPlan: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        mistakes: { type: ["string", "null"] }
        lessons: { type: ["string", "null"] }
        tomorrowGoal: { type: ["string", "null"] }
        isDraft: { type: boolean }
    Rule:
      type: object
      required:
        - id
        - title
        - description
        - category
        - isActive
        - completed
        - sortOrder
        - createdAt
        - updatedAt
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        category: { type: string }
        isActive: { type: boolean }
        completed: { type: boolean }
        sortOrder: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    RuleInput:
      type: object
      required: [title, category]
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string }
        category: { type: string, minLength: 1 }
        isActive: { type: boolean }
        sortOrder: { type: integer }
    RuleUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        description: { type: ["string", "null"] }
        category: { type: string, minLength: 1 }
        isActive: { type: boolean }
        completed: { type: boolean }
        sortOrder: { type: integer }
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
          description: Original file name.
        size:
          type: integer
          minimum: 1
          description: File size in bytes.
        contentType:
          type: string
          minLength: 1
          description: MIME type of the file (e.g. `image/jpeg`).
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
          description: Presigned GCS URL for PUT upload.
        objectPath:
          type: string
          description: Normalized object path (e.g. `/objects/uploads/uuid`). Store this in your database.
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"