# lib/api-zod/src/generated/types/journalInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface JournalInput {
  /** @nullable */
  mood?: number | null;
  /** @nullable */
  confidence?: number | null;
  /** @nullable */
  discipline?: number | null;
  /** @nullable */
  fear?: number | null;
  /** @nullable */
  greed?: number | null;
  /** @nullable */
  focus?: number | null;
  /** @nullable */
  sleep?: number | null;
  /** @nullable */
  tradingPlan?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  mistakes?: string | null;
  /** @nullable */
  lessons?: string | null;
  /** @nullable */
  tomorrowGoal?: string | null;
  isDraft?: boolean;
}