# lib/api-zod/src/generated/types/rule.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Rule {
  id: number;
  title: string;
  /** @nullable */
  description: string | null;
  category: string;
  isActive: boolean;
  completed: boolean;
  sortOrder: number;
  createdAt: Date;
  updatedAt: Date;
}