# lib/api-zod/src/generated/types/ruleInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RuleInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  /** @minLength 1 */
  category: string;
  isActive?: boolean;
  sortOrder?: number;
}